+++ Changelog.md
@@ -1,9 +1,10 @@
 - [ ] Make fixed scalar types default to int32 and int64
-- [ ] Support default values (proto2)
-- [ ] Support required (proto2) fields
 - [ ] Honour [packed=false] flag.
 
 ## 0.9.1: Unreleased
+- [x] Always serialize proto2 type fields
+- [x] Support default values (proto2)
+- [x] Support required (proto2) fields
 - [x] Add options to switch between int64|int32 and int
 - [x] Fix name clash problem with special enum names
 - [x] Refactor serializaton and deserialization to simplify emitted code